+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/resource/cruisecontrol/CruiseControlClientTest.java
@@ -4,7 +4,6 @@
  */
 package io.strimzi.operator.cluster.operator.resource.cruisecontrol;
 
-import io.strimzi.operator.cluster.model.CruiseControl;
 import io.vertx.core.Vertx;
 import io.vertx.junit5.Checkpoint;
 import io.vertx.junit5.VertxExtension;
@@ -74,7 +73,7 @@ public class CruiseControlClientTest {
                 checkpoint.flag();
             })));
 
-        client.setAuthHttpHeader(CruiseControlApiImpl.generateAuthHttpHeader(CruiseControl.API_USER_NAME, CruiseControl.API_USER_NAME));
+        client.setAuthHttpHeader(CruiseControlApiImpl.generateAuthHttpHeader("test", "test"));
         client.getCruiseControlState(HOST, PORT, false)
                 .onComplete(context.succeeding(result -> context.verify(() -> {
                     assertThat(result.getJson().getJsonObject("ExecutorState"),